Move find_empty_sections
