+++ src/database/directory.rs
@@ -20,6 +20,7 @@ impl Directory {
         for (path, doc) in &self.docs {
             let doc_path = relative_path.join(path);
             doc.find_duplicate_sections(&doc_path, issues);
+            doc.find_empty_sections(&doc_path, issues);
         }
         for (path, dir) in &self.dirs {
             dir.check(&relative_path.join(path), issues);
